@@ -1,13 +1,12 @@
 package org.example.dao.service;
 
-import org.example.dao.entity.User;
 import com.baomidou.mybatisplus.extension.service.IService;
-import java.util.List;
+import org.example.dao.entity.User;
 import org.example.domain.dto.UserDTO;
 
 /**
  * <p>
- *  服务类
+ * 服务类
  * </p>
  *
  * @author Xjh
@@ -17,6 +16,7 @@ public interface UserService extends IService<User> {
 
     /**
      * 保存
+     *
      * @param user {@link UserDTO}
      * @return {@link int}
      * @author XJH
@@ -24,14 +24,5 @@ public interface UserService extends IService<User> {
      **/
     boolean insert(UserDTO user);
 
-    /**
-     * 批量保存
-     *
-     * @param dto {@link List<UserDTO>}
-     * @return {@link Integer}
-     * @author Xjh
-     * @since 2023-05-19
-     **/
-    Integer insertBatch(List<UserDTO> dto);
 
 }
