@@ -2,6 +2,7 @@ package org.example.manager.impl;
 
 import cn.hutool.extra.spring.SpringUtil;
 import cn.hutool.json.JSONUtil;
+import jakarta.annotation.Resource;
 import lombok.extern.slf4j.Slf4j;
 import org.example.dao.service.UserService;
 import org.example.domain.dto.UserDTO;
@@ -16,7 +17,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
-import javax.annotation.Resource;
 import java.util.List;
 import java.util.Objects;
 
@@ -48,7 +48,7 @@ public class UserManagerImpl implements UserManager {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void saveBatch(List<UserDTO> users) {
-        userService.insertBatch(users);
+        // userService.insertBatch(users);
     }
 
 
