@@ -1,11 +1,11 @@
 package org.example.config;
 
+import jakarta.annotation.Resource;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
-import javax.annotation.Resource;
 import java.util.concurrent.Future;
 
 @SpringBootTest
@@ -19,7 +19,7 @@ class ExecutorConfigTest {
         executor.execute(() -> System.out.println("大江东去"));
         Future<String> submit = executor.submit(() -> "大江西来");
         System.out.println(submit.get());
-        executor.execute(() -> System.out.println(1/0));
+        executor.execute(() -> System.out.println(1 / 0));
     }
 
 }
